+++ lib/ar_lazy_preload/active_record/base.rb
@@ -10,5 +10,11 @@ module ArLazyPreload
     attr_accessor :lazy_preload_context
 
     delegate :try_preload_lazily, to: :lazy_preload_context, allow_nil: true
+
+    def skip_preload
+      lazy_preload_context&.records&.delete(self)
+      self.lazy_preload_context = nil
+      self
+    end
   end
 end